Split out compiler into separate module, encapsulate
